Don't rely on uninitialized data
When we fail to get file attributes, the file times are left in an
uninitialized state, which may lead to a crash. In particular, this was
showing up in QMessageBox's autotest, where the lastModified time is
being queried on a non-existing file.

Before the refactoring, we were returning a default constructed
QDateTime to queries about different file times, with this change we
will return the time corresponding to a default constructed FILETIME
object.

Reviewed-by: Shane Kearns