Fix a bug that got revealed by 604c51f1fc5c79b7fad12cda911b06b9e6e5005f
The bug has been around for a while, but change 604c51f1fc5c7 made it
emerge.
The problem was that stretches were combined by always *maxing* them.
The values of 'stretch' can be interpreted as this:

-1: (the default) it means that the items should be stretched with
    the stretch factor dervived from the size hints. (In practice this
    means that they are distributed fairly).
 0: Means that the item should not be stretched
>0: Means that the item should be stretch with that number as a factor.

This meant that combining one item with a fixed size(0) and another
item with a default stretch (-1) the combined row stretch would end
up being fixed.

This also fixes how stretches are combined for spanning items too.

Task-number: QTBUG-13551
Reviewed-by: John Tapsell