Fixed QStatusBar not to repaint itself too early
When the toplevel containing a statusbar is initially shown, QStatusBar
will repaint itself way too early - repaint() is called from the
LayoutRequest event causing a whole window repaint before the window was
resized to the proper size.

This caused Qt Creator to paint itself in small size in a maximized
window on its startup.

Task-number: QTBUG-14547
Reviewed-by: Olivier Goffart