Fix a memory leak.
We can't keep strong reference to an engine from a QtData object, because
the object is likely to leak. If it leaks than engine is leaking too which
is really bad.

Persistent::MakeWeak seems to be root of evil. We don't receive callback
when context is about to be destroyed, so we are not able to do proper cleanup.