Clear WSERV content when a native child receives an "expose"
Native child windows on Symbian can cause composition artifacts on S^3
because they own a visible RWindow instance, but the RWindow
technically has no content because child windows paint their content to
their parent's backing store. In the event that a native child widget
completely obscures a top level, the composition engine does not
understand that the child is "empty" and should clear the UI surface
to let the top level "shine through". The result is that articacts can
remain on the UI surface after a transition from a WSERV drawing app to
a Qt (EGL rendering) app.

To fix this issue, we call CWindowGc::Clear() when the native child
widget receives an "expose" event. This clears the UI surface content
and ensures the top level is visible even though it is technically
below another RWindow.

Task-number: QTMOBILITY-484
Reviewed-by: Jani Hautakangas