Fix crash when closing QFontDialog::getFont() dialog
Unfortunately, given the large number of ways of showing a QFontDialog, it's
possible that we try to close the dialog before its event loop is finished.
Setting the Cocoa font panel's isReleaseOnClose property to FALSE ensures the
panel is still around when we effectively exit the dialog's event loop.

Reviewed-by: Richard
Task-number: QTBUG-15666