QScriptQObject: fix fallback in QtDynamicPropertyGetter
QtDynamicPropertyGetter is an accessor and currently in V8 only
interceptors (SetNamedPropertyHandler) can return an empty handle as a
"fallback" -- to let V8 continue to look for the property.

Following Kent suggestion, we simply ask V8 again for the property,
after we deleted our accessor. This is enough for our most basic case,
which is make it fallback to the QObject lazy getter.

It is still in discussion whether we can actually keep those "fast
accessor" for dynamic properties. But for now at least this
"###"-comment is answered and one test is fixed ;-)