Added patch for temporary support for --listgen option in qmake
The --listgen option is needed in qmake for Symbian platform builds
to list the files generated by qmake. This patch enables qmake to
recognize this option, but is not functional, just to allow builds to
succeed. This patch is not applied to qmake directly, but only made
available in the source tree for usage in the build setup.

Reviewed-by: TrustMe