Fix the test since JS property have higher priority than Dynamic property
I was mistaken :-P. The JS property should not be hidden when there's a
dynamic property in the object. So setting a previously created js property
shouldn't affect the dynamic property.

Renamed the test so it now tells us that "dynamic property do not hide
js property". This behavior is consistent with Qt 4.5.3 (the "classic"
engine).

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>