Cocoa: add support for QEventLoop::ExcludeUserInputEvents
Seems like this was just missing when used together with the
DialogExec/EventLoopExec flags. This patch avoid using cocoas
own run methods for dispatching events when we want to
exclude user input events, and instead do it ourselves.

This patch will only fix this issue when no modal windows are
showing. For modal windows, a bit more research is needed.