Fix KERN-EXEC 0 panic on exit when bearer is searching for WLANs
The access point scanner cancels itself in the destructor.
This requires the handle to be valid, but it was closed in the symbian
engine destructor immediately before deleting the AP scanner.
Because of the way symbian active objects work, the crashing function
is only called if there was an asynchronous request in progress.
So it could be missed in cases where the scan completes faster than the
test case.

Task-number: QTBUG-16484
Reviewed-by: Markus Goetz