Fix handle leak in symbian QTimer implementation
The timer handle was only being closed when a timer was cancelled,
which resulted in a leak for one shot timers that have completed normally.

Instead the timer is now closed in a destructor (closing null handles is
safe, so it doesn't matter if the handle was never created - e.g. in the
case of a zero timer)
Also added a handle check before creating a timer to prevent a leak in
case the start function is called twice in the backend.

Task-number: QTBUG-16380
Reviewed-by: mread
(cherry picked from commit 2b1b617664bfc78f6e95e53dc0f9749bd1f2d27a)