set() and setProperty() should not always trigger change signals
Fix set() and setProperty() to only trigger itemsChanged(), and the
the <property>Changed signals for items returned by get(), if the
new value is different from the old one. The exception to this are
list values, as it is inefficient to check the sublists and also the
model classes are due to be revised.

This also fixes models used with a WorkerScript to emit
itemsChanged() with the correct roles.

Task-number: QTBUG-14620
Reviewed-by: Michael Brasser