Fix a memory leak when Persistent::MakeWeak is used.
As weak reference doesn't work as supposed in v8 (no callbacks are
called if a context is about to be destroyed), we have to track
allocated resources on our own. It is done in QScriptEnginePrivate
by 3 functions:
- registerAdditionalResources(QtDataBase *data)
- unregisterAdditionalResources(QtDataBase *data)
- deallocateAdditionalResources();

New class QtDataBase was introduced, it privides a common path for
QtData instances destruction by adding a virtual destructor. The class
is not template based so a common pointer type can be used in
QScriptEnginePrivate::deallocateAdditionalResources.

Additionally all QNativeFunction classes were moved to the QtDataBase
hierarchy so they can profit from resource tracking feature.