Allow IAP to be selected in Phonon MMF backend
The Phonon API does not provide any way for the client to specify
which network connection should be used for streaming playback.
If the application already has a connection open, using a bearer
other than the default (e.g. using WiFi when the device default is
GPRS), it may be desirable to use it for streaming, rather than
allowing the Phonon backend to open a second connection on the
default bearer.

This patch adds a custom property on the Phonon::MediaObject,
called InternetAccessPointName.  The client can specify the IAP
which Phonon should use by setting this property.

Note that support for this property is only provided in the Phonon
MMF backend.

Task-number: QTBUG-11436
Reviewed-by: Gareth Stockwell