Fix the warnings when QBasicTimer are member of the ObjectPrivate
If a class has QBasicTimer as member of a derived class QObjectPrivate,
the QBasicTimer will be destroyed after the timer has been cleaned.

The QBasicTimer will stop the timer in its destructor. But that means
that the same timer will be stoped twice, leading to ugly bugs.

Reviewed-by: Thiago
Reviewed-by: Brad
Task-number: QTBUG-16558
Task-number: QTBUG-16175