v8: We don't need Qt for building libv8
This link was causing problems when running mksnapshot during the
building process, since it used libv8 and wanted libQtGui and
libQtCore.

I think this issue was hidden for so long by the fact that if you have
those libs available in LD_LIBRARY_PATH, that would be enough for
mksnapshot to run -- since it doesn't use Qt symbols at all.