QTBUG-16110 updating the softkeys when app comes to foreground
When an app is in the background and creates a window with softkeys,
the softkeys are not updated when the app comes to the foreground, so
they look faded. This change updates the softkeys when the app comes
to the forground.

Task-number: QTBUG-16110
Reviewed-by: Jason Barron
(cherry picked from commit 37faa3ba7c66d14e1ecbf75e93d9f0d27fa57a76)