Fix to data detaching in OpenVG pixmap data
QImage::converToFormat(format) won't detach QImage if
format stays the same. Because of this, internal QImage
buffer of QVGPixmapData was not deep copied. Drawing
to QPixmap target modified the original QImage that was
used to create QPixmap and in some cases crashed because
paint engine referred to wrong target.

Task-number: QT-4407
Reviewed-by: Jason Barron