Fix some issues with socket handle allocation
QAbstractSocket requires a handle of -1 to be returned when accept fails.
This is a common case, as it calls accept() in a loop to get all incoming
connections. Also the blank socket wasn't being closed properly on failure
so that is also fixed.

Fixed a possible race condition in QSymbianSocketEngine::close - the
socket descriptor is now deregistered before closing the symbian socket
to avoid another thread getting the symbian handle reused in open and
asserting when trying to register it.

This patch also adds debug around socket handle allocation when socket
engine debug is enabled.

Reviewed-by: Markus Goetz