QScriptContext: 'arguments' in JS for native calls
Makes the 'arguments' property available in the JS evaluation
context. Its object is the same as the one returned by
argumentsObject().

We use a lazy approach, instead of setting the property directly, we
create an accessor for the first run. So the penalty is only setting
up the accessor in the activationObject ("extension object" in v8
lingo).

Reviewed-by: Olivier Goffart