QScriptContext: lazy initialize 'arguments' property for native calls
Now QScriptEngine::evaluate() takes care of initializing the arguments
property. So "pure" C++ functions won't pay for what they don't use.

When using evaluate() in a native function context, it'll initialize
the 'arguments' property, either with the object if it already exists
or with an accessor that'll will delay it's creation to the first use.

Reviewed-by: Olivier Goffart