Added Symbian deployment localization for makefile build system.
This is a complement to the 16575f7aef840b6aae0dc767468ab713fbcfd7a6
commit, which adds localization based on TRANSLATIONS keywords for
Raptor and abld.

In addition, since the __PRODUCT_INCLUDE__ define was creating a lot
of trouble regarding < and >, it was refactored into its own source
file, which is automatically included before every source file.

Task:     QTBUG-15292
RevBy:    Miikka Heikkinen