Improved QMAKE_POST_LINK support in symbian-sbsv2
QMAKE_POST_LINK value handling now uses the same automatic replacements
as QMAKE_EXTRA_COMPILERS and QMAKE_EXTRA_TARGETS handling.
In practice this means that it is now possible to use $$QMAKE_COPY and
friends in QMAKE_POST_LINK value also with symbian-sbsv2 mkspec, and
that any backslashes in the value are assumed to be path separators and
will be converted to forward slashes.

Task-number: QTBUG-16753
Reviewed-by: Janne Koskinen