Ongoing improvements and fixes to syntax highlighting.
Added an atom for JavaScript code.
Fixed example file quoting to use the appropriate atom for each file.
Changed C++ code marking to mark up classes as types.
Fixed C++ quoting bug that caused the last character to be lost.
Fixed indentation of code to not insert spaces inside elements.
Documented the change in the meaning of the \c command.
Simplified the use of CSS classes in the HTML output.