Cocoa/Alien: bugfix scrolling, dirty region issue
When scrolling a widget we also need to scroll the already
marked dirty regions on the widget along with the scroll.
Since we might end up scrolling several time before a drawRect
is called (normal with inertia scrolling), we need to mark exposed
areas during scrolling as dirty as well, so they get handled
correctly on subsequent scrolls.