Fix crashes in httpProxyCommandsSynchronous test case
Due to a race condition, the test could fail before its QThread had even
started, resulting in a crash during thread startup. This is addressed
using a semaphore to synchronise startup & ensure the sub thread is running
before continuing the test.
Additionally, if the test failed the QThread would be destroyed while
still running, which according to the documentation will cause crashes.
This is addressed using QScopedPointer to clean up when the test exits
early due to failure.

Reviewed-by: Markus Goetz