Use the global message pool to avoid KErrServerBusy errors
Symbian IPC has two modes, either a preallocated message pool for the
process (up to 255 allowed), or the global message pool.
The default for the socket server is 8 preallocated slots, which is not
enough to run the qnetworkreply autotest. However preallocating a large
number of slots would be wasteful for an application that only uses one
or two sockets at a time.
Since we can't predict the usage, the global pool seems the best option.

Reviewed-by: Iain