Add support for disabling touch on to enhance scrolling in Cocoa
On Mac/Cocoa, enabling touch on a widget slows down the scrolling
performance of the whole application. It seems Cocoa spends some
time in the background figuring out what to do with the touch
events, and whether or not it should convert them to scroll/wheel
events. Therefore, it makes sense to no subscribe for touch when
the mouse is not over the widget, This patch implements that
strategy, and the effect is huge when tested agains creator.

Rev-By: brad