Fixed overshooting with a very low start velocity.
Before this patch, the slow down (overshoot) was a separate scroll segment.
This could lead to an actual acceleration when going into overshoot with a
very low velocity.  The new approach is to just continue with the normal
scroll segment, until the max overshoot distance is reached and only add
one additional segment for the bounce-back animation to complete the
overshoot animation.

Reviewed-by: Ralf Engels