qnetworkreply test stability on symbian
The first few tests could randomly timeout due to delays in starting
the WLAN connection.
This is addressed by opening the default bearer in initTestCase, and
closing it in cleanupTestCase. This will keep the connection active for
the duration of the test, instead of activating/deactivating continuously.

The first SSL test could randomly timeout due to the delay in fetching the
CA certificates from the system (which can be very slow on some phones /
SIM cards).
This is addressed by preloading them in initTestCase - all the tests will
then get system SSL certs from the cache.

Reviewed-by: Markus Goetz