interpret backslash escaping of quotes in split_arg_list()
split_value_list() interprets them, too, but this is useless if the
higher layer doesn't and is thus confused by unmatched quotes.

note the top-level parsing layer doesn't support escapes, but that's ok,
because there function calls are "isolated" by non-escapable parenthesis
counting (and the RHS of assignments is not subject to any parsing at
all).

Reviewed-by: mariusSO