Orientation control implementation for Symbian
This used the orientation control QWidget attributes API from
maemo5, and provides a simple implementation for Symbian.

The essense of the implementation is that the latest setting
of one of these QWidget orientation attributes will set the
orientation for the whole app. Testing the attributes will return
only the last attribute set, it will not return the app orientation
state.

A new task, QTBUG-16972, has been created to provide a more
comprehensive implementation in the future. This may provide a more
effective emulation of the maemo5 behaviour, or may incorporate
further reaching concepts for QML rotations.

Task-number: QTBUG-11785
Reviewed-by: Shane Kearns