Make sure QStateMachine stops when it's told to
If QStateMachine::stop() was called by an event test or while
transitioning to another state, the state machine could end up
in an inconsistent internal state (stop==true, but the stopped()
signal was not emitted). This also caused the machine to behave
incorrectly if it was then restarted (it would immediately stop
upon receiving the first event).

Solution: Move the stop-handling after the event processing loop,
so that it always takes precedence over other possible reasons
for exiting the loop (event queues exhausted, final state
entered).

Task-number: QTBUG-16463
Reviewed-by: Eskil Abrahamsen Blomfeldt