QComboBox popup incorrectly opened in Sym^3
QCombobox popup opens into incorrect position in Sym^3. This is due
to that there is a new layout in use, where softkeys are positioned
at the bottom. The current implementation just checks if native stacon
component is in use and if not, puts the popup to left/right border
of the application area (depending on UI direction).

Task-number: QTBUG-16886
Reviewed-by: Miikka Heikkinen