Fix objectDeleted test.
When wrapped qobject is about to be destroyed all connections between
that object and others should be destroyed too.

Using QScriptValue which is an empty qobject wrapper, cause in most
cases an exception.

This patch fixes objectDeleted test, I've added some expected fail,
because error message is missing a property name. That can be easily
fixed as soon as v8 bug 1072 is fixed.