Replace the handcoded math and change the timer buckets again.
Reduce from 8 to 6 buckets and increase the step between each
bucket. This way, the second bucket is now of 224 timers, which should
be enough for 99.9% of the applications.

Also change the hardcoded math to calculations using enum values. This
helps in changing the timer buckets again in the future.

Also fix the last bucket not to have a timer ID of 16777216, as that
is not valid.

Reviewed-by: Olivier Goffart