No longer replace dash and dot in TARGET with underscore in Symbian
There is no fundamental reason to not have dash or dot in binary names
in Symbian, so do not replace them with underscore.

One thing that doesn't work with a dot in the filename is launching
an application via resources, so automatic resource generation is
suppressed for applications that have a dot in filename portion of
the TARGET value.

Task-number: QTBUG-16888
Reviewed-by: axis
(cherry picked from commit 4ec245a3e75470186557d00b2383af3872a720b0)