Restore Qt 4.6 behaviour: exec() always enters the event loop.
In Qt 4.6 as well as 4.7's QCoreApplication and QEventLoop, calling
exec() always enters the event loop, even if you had tried to
quit()/exit() it before entering, with one exception (noted in the
unit tests; this difference has been in Qt since at least Qt 4.2).

Add unit tests to ensure all of the three classes have the same
behaviour.

Decide if we want to match the behaviours in Qt 4.8.

Reviewed-by: Bradley T. Hughes