Fix crash in context2d during garbage collection.
The crash was introduced together with memory management change and it
was cause by missing resource unregistration. Missing call was in a
template code, unlucky it was not possible to call function
unregisterAdditionalResources directly as it was really difficult to
check if / were object was registered before.

I solved the issue by
- changing ~QtDataBase so now it is responsible for calling the
unregister function
- allowing to call QBagContainer::remove for an object that is not
associated with any container.

The code is a bit redundant, in some cases an object can call unregister
function twice.