Cocoa: fix autotest, qgraphicsview::moveItemWhileScrolling
This test reveals a real bug when updating and scrolling
at the same time in graphicsview. The problem is that the scroll
will issue an update to cocoa that gets handled before already
queued Qt events. This means that the viewport of graphicsview
will receive a paintevent, and clear dirtyScrollOffset, before
the posted call to _q_processDirtyItems is executed. This means
that we 'loose' the scrolling information upon painting.

I discussed this problem with Bjørn Erik, and he had strong opinions
about this promblem not worthwhile fixing. The raster engine is
anyway about to be default on Mac, and there the problem will not exist.