Fix alignment issue causing crash in QtScript/JavaScriptCore
When creating a substring, JSC::UStringImpl required that the
base string pointer was 8-byte aligned. However, on platforms
where FastMalloc isn't enabled (such as Symbian), it's possible
that the system malloc() returns a pointer that is only 4-byte
aligned. (On Symbian, this can happen if the argument to malloc()
itself isn't a multiple of 8.)

Cherry-picked http://trac.webkit.org/changeset/54743 from WebKit
trunk, which fixes this issue. (The commit happened shortly after
we rebased QtScript/JSC for 4.7, so it applies cleanly to our copy.)

Task-number: QTBUG-16828
Reviewed-by: Simon Hausmann