Allow functions to be passed in as values for grouped properties
Using QDeclarativeProperty::write() works for grouped properties but
stops functions values from being passed in, and using
QScriptValue::setProperty() on the object being created allows functions
to be passed in but doesn't work for grouped properties. This fix
walks down the tree for grouped properties to find the property that
should be set so that functions can be set for grouped property values.

Reviewed-by: Aaron Kennedy