Cocoa: fix qgraphicsview autotest
Moving the cursor to the corner of the desktop (0, 0) will typically
trigger things to happend in the OS, at least on Mac (spaces, synergy etc).
So I changed those tests to move it to (1, 1) instead. The reason to move
it to (0, 0) was just to get the mouse away from the window anyway.

And second, it takes some time after using QCursor::setPos on Mac
before the events from the move actually arrive. And many tests
that use this function relies on this. So I introduces a little
delay. This alone will fix a couble of tests.