Invalidate QScriptPrograms when engine is destroyed
If the engine is destroyed before the program, the program
must be invalidated; otherwise the program destructor will
access a stale engine pointer, which can cause a crash
(it crashes on Symbian, but "only" gives a Valgrind warning
on Linux for our autotests).

We need to keep track of all associated programs, just like
we already do for values and strings. This fix follows the
exact same pattern, but uses a QSet to keep the patch minimal.

No new tests, but the evaluateProgram() test runs successfully
on Symbian now, and there are no more Valgrind warnings.

Task-number: QTBUG-16987
Reviewed-by: Olivier Goffart