Cocoa/Alien: fix qgraphicseffectsource autotest
The test discovered a serious bug where we for alien widgets on mac
returned a zero sized paintdevice. The reason is that we used to
ask the nsview backing the widget for its size, but with alien, there
where no such view. Very strange that this was not seen visually for
any of the many applications tested. This patch does the same
implementation strategy as found in QWidget::metric in qwidget_x11.cpp