QNAM: Threaded HTTP implementation
HTTP requests are run in a separate thread now.
This required some big changes in the QNetworkAccessHttpBackend.
There is a new class QHttpThreadDelegate which lives in the
HTTP thread and is the communication layer between HTTP code
and QNetworkAccessHttpBackend. Communication is done
via signals/slots.
The synchronous HTTP code (private QtWebKit API) also had to
be completely re-worked and uses its own thread now.

Reviewed-by: Peter Hartmann
Task-number: QTBUG-14162