Fix for pendingDatagramSize() in symbian socket engine
The low level socket option returns the size including IP headers, as
there is an option in receive to include headers in the datagram.
This is OK for a "size will not exceed" metric for buffer allocation,
but Qt relies on it being an accurate size.
Open C did this by subtracting 28, but that isn't valid for IPv6
which has a 40 byte header. (we can't tell whether the buffered datagram
was received over IPv4 or IPv6)

To fix this, do a read with the peek option set, and only care about the
size. In future it would be good to not peek, but rather return this buffer
to a following call to readDatagram.

Reviewed-by: Aaron Tunney