Fix low level broadcast test on symbian
Broadcast and multicast are optional features for network interfaces on
Symbian. WLAN supports both, while IPv4 and IPv6 loopback support only
multicast.

The test tries to start the bearer, to get an interface that does support
broadcast.
There is an expect fail for the 0 bytes written case, as this is what
happens on symbian if none of the interfaces support broadcast.
On the phones which don't have WLAN, this is likely.

Reviewed-by: Markus Goetz