Idle detector thread improvements
The idle detector thread was creating its own heap, which was rather
unnnecesary as it doesn't actually allocate any objects. Now it shares
the main thread's heap.

It was also unnamed, which makes life harder when there is a crash.
The debugger is given information about all threads in a crash. When a
thread is unnamed, you can't easily tell if its relevant to the crash
or not. Now it is called "IdleDetectorThread".

Task-number: QTBUG-17073
Reviewed-by: axis