Fix QMutex can deadlock when calling tryLock
in the unix code, if the QMutexPrivate::wait() with a timeout
expires in the same moment that the mutex is released,  wakeup
would be set, but would be then ignored. (reset to false
quickly after)

If we waken up between the timeout and the re-aquisition of
the internal mutex, we consider that the mutex has been locked.

Reviewed-by: brad
Task-number: QTBUG-16115