Fixed a bug where the softkeys would leave a see-through "hole".
In cases where the widget was created from the event loop instead of
main(), the middle three softkeys would not get the right visibility
and would leave a "hole" in the application where one could see what
was beneath it. It could also be clicked, which would bring the app
to the background which was really annoying.

Fixed by setting the visibility manually upon creation. It should
still be possible to alter the visibility of the toolbar manually
after the widget creation, if someone requires that.

Task:     QTBUG-16578
RevBy:    Gareth Stockwell
AutoTest: N/A, but tested with various combinations of fullscreen,
          non-fullscreen and softkey visibility modes. No
          regressions.
(cherry picked from commit 231dcfc8b6a90c3a0ab4869e0a9ab45403fd9a8c)