Using a better flag to control the fast allocator
The Qt fast allocator should be used on any Symbian platform where it
is not supplied by euser. But it's better if the code is not there when
euser provides the fast allocator as euser's implementation will be
better maintained.

The previous flag SYMBIAN_GRAPHICS_WSERV_QT_EFFECTS was only associated
with euser's fast allocator in S60 10.1. __SYMBIAN_KERNEL_HYBRID_HEAP__
is more accurately associated with it in the new MCL.

Task-number: QTBUG-15901
Reviewed-by: Miikka Heikkinen