Improve performance of partial updates in raster window surface on X11.
An XSync is needed to prevent the raster engine from writing to the
shared memory raster backbuffer at the same time the X server reads from
it. However, instead of doing the sync right after telling X to blit, we
can do it right before we start writing to it with raster. At this point
there will on average be less processing heavy X commands in the command
queue, and we thus spend less time blocking the GUI thread.

Measured frame rate improvement of 20 - 60 % in an update stress testing
the performance of partial updates.

Idea-from: Olivier Goffart
Reviewed-by: Bjørn Erik Nilsen