Don't crash when BMP color table is broken
If the BMP's number of color table entries is out of bounds, we would
resize our color table vector to a silly value, leading to crashes
later on. If the number of color table entries is larger than 256, just
stop processing the BMP since it's most probably corrupt.

Task-number: QT-4534
Reviewed-by: Robert Griebl