Fixed mapFromGlobal in some window managers on X11
The previous mapFromGlobal optimization in
cdd776a91e65bf5c30cea1bab9823134a3f797d0 broke behavior in some window
managers like metacity when resizing by dragging the top-left corner.

Most likely it is a bad behaviour of the window manager (according to
ICCCM 4.1.5 we should get a ConfigureNotify event with x and y in
root coordinates if you consider this is a "movement" of a window).

Reviewed-by: Olivier Goffart