Check that _POSIX_THREAD_SAFE_FUNCTIONS is larger than zero.
POSIX allows this symbol to be defined to -1 to indicate that the
interfaces aren't present. On FreeBSD 8, it's defined to -1, but the
functions are present (supposedly, they don't work -- just stubs).

Guessing that OpenBSD is the same.

Also removing the QT_NO_THREAD, since Qt can't be built without
threading support. It's only used by our bootstrapped tools, which in
turn use QCoreApplication only (no QApplication).

Reviewed-By: Bradley T. Hughes