Call onAdd() for first items added to ListView and GridView
The bug was caused by the fact that the itemsInserted()
implementation in ListView and GridView returned early for cases
where (model->count() <= 1) and so did not call onAdd() unlike for
other cases. This change also fixes some incorrect values in the
header() and footer() unit tests for ListView and GridView.

Change-Id: I24f5c86afcd62c26e17b0932f257f4767a287b8e
Task-number: QTBUG-15642
Reviewed-by: Martin Jones