Fix auto test failure.
Allow the input context to be reset twice.  The context is reset
because of two events, the first is because an item lost focus and that
reset needs to occur before the focus moves away from the item or events
won't be delivered to it.  The second redundant reset is because the
Qt::WA_InputMethodEnabled flag was cleared on the view widget because
no item has focus and it has no way to know the context doesn't need
to be reset.

Change-Id: Ie3b3fc6898d144ed3f8b3822e49ea0eee7e029f4
Reviewed-by: Martin Jones